Cellula_2week_Omar_Arnous/requirements.txt
transformers
torch
pandas
streamlit
Pillow
matplotlib
dotenv